Allow layer definitions to be processed as templates
This commit adds the ability to process the layer definitions as Jinja2
templates just like any other Klipper macro. All of the normal objects
available to macros are also available to layers.

As an example, the following layer definition:

    [led_effect test1]
    autostart: false
    heater: extruder
    leds:
      neopixel:case_lights
    layers:
      {% set min_temp = printer.configfile.settings.extruder.min_extrude_temp %}
      heater {min_temp} 0 add (1,0,0)

will result in the following layer being defined:

    heater 170.0 0 add (1,0,0)

Layers are then processed as normal.

It is also possible to have layers be re-evaluated when an effect is
enabled. This has the benefit of using the current printer state.
However, it will use CPU cycles to re-compute the layers.

Signed-off-by: Mitko Haralanov <voidtrance@gmail.com>