Allow passing of effect parameters with SET_LED_EFFECT
This commit adds the ability to pass parameters to the LED effects
through the SET_LED_EFFECT command.

If an effect is defined as `dynamic` (which re-evaluates the layers
template each time the event is activated), the SET_LED_EFFECT
command can now pass parameters to the template just like GCode
macros can.

This mechanism allow for the effect to be changed dynamically. In
fact, it could provide a way for definiting dynamic effects.

Signed-off-by: Mitko Haralanov <voidtrance@gmail.com>